// packages/editor-ui/src/Interface.ts
export interface IUpdateInformation {
	name: string;
	key?: string;
	value:
		| string
		| number
		| { [key: string]: string | number | boolean }
		| NodeParameterValueType
		| INodeParameters; // with null makes problems in NodeSettings.vue
	node?: string;
	oldValue?: string | number;
}
// ...
export interface IRestApiContext {
	baseUrl: string;
	sessionId: string;
}
// ...
export interface ExternalSecretsProvider {
	id: string;
	name: string;
	connected: boolean;
	connectedAt?: string;
	properties: INodeProperties[];
	data: Record<string, IUpdateInformation['value']>;
}